+++ apps/boruta_admin/assets/src/views/Layouts/Main.vue
@@ -147,19 +147,22 @@
       <a @click="toggleDarkMode()" class="dark-mode">
         <i class="sun icon" :class="{ 'outline': currentMode }"></i>
       </a>
-      &copy; 2024 malachit
+      <Feedback />
+      &copy; 2025 malachit
     </footer>
   </div>
 </template>
 
 <script>
 import Header from '../../components/Header.vue'
+import Feedback from '../../components/Feedback.vue'
 import Breadcrumb from '../../components/Breadcrumb.vue'
 
 export default {
   name: 'Main',
   components: {
     Header,
+    Feedback,
     Breadcrumb
   },
   data () {
@@ -236,6 +239,7 @@ export default {
     cursor: default;
   }
   footer {
+    position: relative;
     z-index: 200;
     text-align: right;
     padding: 1rem;