+++ apps/boruta_identity/lib/boruta_identity_web/views/template_view.ex
@@ -143,17 +143,22 @@ defmodule BorutaIdentityWeb.TemplateView do
     |> context(Map.delete(assigns, :scopes))
   end
 
+  def context(context, %{code: code} = assigns) do
+
+    %{code: code}
+    |> Map.merge(context)
+    |> context(Map.delete(assigns, :code))
+  end
+
   def context(context, %{}), do: context
 
   defp text_from_credential_offer(credential_offer) do
     # TODO Jason.Encode implementation for CredentialOfferResponse
-    "openid-credential-offer://?credential_offer=#{
-      credential_offer
+    "openid-credential-offer://?credential_offer=#{credential_offer
       |> Map.from_struct()
-      |> Map.delete(:credentials)
+      |> Map.take([:credential_configuration_ids, :credential_issuer, :grants])
       |> Jason.encode!()
-      |> URI.encode_www_form()
-    }"
+      |> URI.encode_www_form()}"
   end
 
   defp paths(conn, assigns) do
@@ -185,6 +190,8 @@ defmodule BorutaIdentityWeb.TemplateView do
         Routes.user_session_path(BorutaIdentityWeb.Endpoint, :delete, %{request: request}),
       edit_user_path:
         Routes.user_settings_path(BorutaIdentityWeb.Endpoint, :edit, %{request: request}),
+      destroy_user_path:
+        Routes.user_settings_path(BorutaIdentityWeb.Endpoint, :destroy, %{request: request}),
       new_user_totp_registration_path:
         Routes.totp_path(BorutaIdentityWeb.Endpoint, :new, %{request: request}),
       create_user_totp_registration_path: