+++ apps/boruta_identity/priv/repo/migrations/20241130000259_add_check_password_to_identity_providers.exs
@@ -1,0 +1,9 @@
+defmodule BorutaIdentity.Repo.Migrations.AddCheckPasswordToIdentityProviders do
+  use Ecto.Migration
+
+  def change do
+    alter table(:identity_providers) do
+      add :check_password, :boolean, null: false, default: true
+    end
+  end
+end