+++ apps/boruta_identity/test/boruta_identity/admin_test.exs
@@ -101,10 +101,35 @@ defmodule BorutaIdentity.AdminTest do
     end
 
     test "returns deleted user" do
-      %User{id: user_id, uid: user_uid} = user_fixture()
+      %User{id: user_id, uid: user_uid} = user_fixture(%{}, "internal")
       assert {:ok, %User{id: ^user_id}} = Admin.delete_user(user_id)
       refute Repo.get(User, user_id)
       refute Repo.get(Internal.User, user_uid)
+    end
+  end
+
+  describe "delete_user/1 with federated backend" do
+    test "returns an error" do
+      assert Admin.delete_user(Ecto.UUID.generate()) == {:error, :not_found}
+    end
+
+    test "returns deleted user" do
+      %User{id: user_id} = user_fixture(%{}, "federated")
+      assert {:ok, %User{id: ^user_id}} = Admin.delete_user(user_id)
+      refute Repo.get(User, user_id)
+    end
+  end
+
+
+  describe "delete_user/1 with wallet backend" do
+    test "returns an error" do
+      assert Admin.delete_user("did:key:test") == {:error, :not_found}
+    end
+
+    test "returns deleted user" do
+      %User{id: user_id} = user_fixture(%{}, "wallet")
+      assert {:ok, %User{id: ^user_id}} = Admin.delete_user(user_id)
+      refute Repo.get(User, user_id)
     end
   end
 
@@ -146,13 +171,52 @@ defmodule BorutaIdentity.AdminTest do
       params = %{
         username: "test@created.email",
         password: "a valid password",
-        metadata: %{"attribute_test" => "attribute_test value"}
+        metadata: %{
+          "attribute_test" => %{
+            "value" => "attribute_test value",
+            "status" => "valid"
+          }
+        }
       }
 
       assert {:ok,
               %User{
-                metadata: %{"attribute_test" => "attribute_test value"}
+                metadata: %{
+                  "attribute_test" => %{
+                    "value" => "attribute_test value",
+                    "status" => "valid"
+                  }
+                }
               }} = Admin.create_user(backend, params)
+    end
+
+    test "returns an error with invalid metadata", %{backend: backend} do
+      metadata_field = %{
+        "attribute_name" => "attribute_test"
+      }
+
+      {:ok, backend} =
+        Ecto.Changeset.change(backend, %{
+          metadata_fields: [
+            metadata_field
+          ]
+        })
+        |> Repo.update()
+
+      params = %{
+        username: "test@created.email",
+        password: "a valid password",
+        metadata: %{
+          "attribute_test" => %{
+            "value" => "attribute_test value"
+          }
+        }
+      }
+
+      assert Enum.empty?(Repo.all(Internal.User))
+      assert_raise Ecto.InvalidChangesetError, fn ->
+        Admin.create_user(backend, params) == {:error, %Ecto.Changeset{}}
+      end
     end
 
     test "creates a user with a group", %{backend: backend} do
@@ -227,10 +291,35 @@ defmodule BorutaIdentity.AdminTest do
         Ecto.Changeset.change(user.backend, %{metadata_fields: [%{attribute_name: "test"}]})
         |> Repo.update()
 
-      metadata = %{"test" => "test value"}
+      metadata = %{
+        "attribute_test" => %{
+          "value" => "attribute_test value",
+          "status" => "valid"
+        }
+      }
+
       user_params = %{metadata: metadata}
 
       assert {:ok, %User{metadata: ^metadata}} = Admin.update_user(user, user_params)
+    end
+
+    test "returns an error with invalid metadata", %{user: user} do
+      {:ok, _backend} =
+        Ecto.Changeset.change(user.backend, %{metadata_fields: [%{attribute_name: "test"}]})
+        |> Repo.update()
+
+      metadata = %{
+        "attribute_test" => %{
+          "value" => "attribute_test value"
+        }
+      }
+
+      user_params = %{metadata: metadata}
+
+      assert {:error,
+              %Ecto.Changeset{
+                errors: [metadata: {"Required property status was not present. at #", []}]
+              }} = Admin.update_user(user, user_params)
     end
 
     test "returns an error if group is not unique", %{user: user} do